+++ CHANGELOG.md
@@ -6,7 +6,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Updated
+
+- VTEX-Tachyons version
+
 ### Fixed
+
 - Spinner proptypes
 
 ## [5.6.2] - 2018-8-17