+++ CHANGELOG.md
@@ -24,6 +24,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - **Input** `onKeyUp` and `onKeyDown` hooks
 - **Input** `groupBottom` prop
 
+### Changed
+
+- **Alert** Add semantic colors and remove default Alert
+
 ## [5.5.1] - 2018-08-13
 
 ## [5.5.0] - 2018-08-13